Rename files implementing utility functions for building station list
Also rename directory containing corresponding test data.
