+++ handlers/data_access.js
@@ -10,7 +10,7 @@ const csv = require('dwd-csv-helper')
 const _ = require('lodash')
 const { convertUnit } = require('../lib/unit_conversion.js')
 const gf = require('../lib/grib_functions')
-const su = require('../lib/station_utils.js')
+const su = require('../lib/stations_utils.js')
 const mvu = require('../lib/measured_values_utils.js')
 const reqU = require('../lib/request_utils.js')
 const ru = require('../lib/response_utils.js')
