+++ handlers/measurements.js
@@ -7,7 +7,7 @@ const _ = require('lodash')
 const path = require('path')
 const moment = require('moment')
 const fs = require('fs-extra')
-const su = require('../lib/station_utils')
+const su = require('../lib/stations_utils')
 const assert = require('assert')
 const processenv = require('processenv')
 var bunyan = require('bunyan')
