+++ lib/forecast_utils.js
@@ -8,7 +8,7 @@ const moment = require('moment')
 const path = require('path')
 const _ = require('lodash')
 const gu = require('../lib/general_utils.js')
-const su = require('../lib/station_utils.js')
+const su = require('../lib/stations_utils.js')
 const gf = require('../lib/grib_functions')
 const bunyan = require('bunyan')
 const log = bunyan.createLogger({
